refactor: split runtime.rs into sub-modules
Split the 1,319-line src/runtime.rs into src/runtime/ directory with four
sub-modules, each under 400 impl lines:

- types.rs (107 impl lines): Core enums/structs — RuntimeTopology,
  RuntimeState, RuntimeNodeId, RuntimeAddress, BaseTypeRegistry,
  RuntimeRequest, SessionOutcome
- traits.rs (123 impl lines): Trait definitions and driver implementations —
  RuntimeTopologyDriver, RuntimeMailboxTransport, RuntimeSupervisor,
  InProcessTopologyDriver, LoopbackMeshTopologyDriver, etc.
- session.rs (299 impl lines): RuntimeKernel session execution pipeline —
  execute_session, persist_session_scratch, build_reflection,
  snapshot_for, ReflectiveRuntime impl
- mod.rs (383 impl lines): RuntimePorts, RuntimeKernel struct/compose/
  lifecycle/subordinate management, re-exports preserving public API

All public API paths via crate::runtime::* are preserved through re-exports
in mod.rs. Tests moved to their appropriate sub-modules.
